requirements.txt
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
Flask==2.3.2
google-api-core==2.11.1
google-api-python-client==2.96.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.60.0
gunicorn==20.0.4
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.25.2
oauthlib==3.2.2
packaging==23.1
pandas==2.0.3
protobuf==4.24.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pygsheets==2.0.6
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
Werkzeug==2.3.6
zipp==3.16.2
openai==0.28.1
line-bot-sdk==3.1.0
